Add files via upload
Exploratory Data Analysis (EDA)
This project performs an Exploratory Data Analysis (EDA) on the classic Iris dataset. The goal is to gain a better understanding of the characteristics of the three iris species (Setosa, Versicolor, and Virginica) through statistical analysis, visualizations, and examination of variable correlations. Tools such as Python, Pandas, Matplotlib, and Seaborn are used to uncover meaningful patterns and relationships within the data.